bgp: align ADD-PATH test with EcmpGroup.nh_count being u8
Two clippy/test gate failures from CI:

The IpAddr import was left over from the original test draft that
constructed IpAddr::V4(...) for next-hop arguments. The current
shape passes Ipv4Addr directly to the wire-format helpers and
never names the IpAddr type, so the import is dead under
-D unused-imports.

EcmpGroup.nh_count is u8 (see crates/modules/fast-path/src/fib/
types.rs); the polling loop declared its observation variable as
u32 to give the error message a wider count. Rust does not coerce
u8 to u32 across an `=` assignment without an explicit cast, so
the binding now matches the field's type.

+++ crates/modules/fast-path/tests/route_source_bgp_addpath.rs
@@ -16,7 +16,7 @@
 
 #![cfg(target_os = "linux")]
 
-use std::net::{IpAddr, Ipv4Addr, SocketAddr};
+use std::net::{Ipv4Addr, SocketAddr};
 use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Once;
@@ -352,7 +352,7 @@ fn add_path_two_updates_yield_ecmp_via_bgp_listener() {
         let key = LpmKey::new(24, [192, 0, 2, 0]);
         let deadline = Instant::now() + Duration::from_secs(5);
         let mut last_seen: Option<FibValue> = None;
-        let mut last_count: u32 = 0;
+        let mut last_count: u8 = 0;
         loop {
             // Fresh handle each poll so we observe the programmer's
             // writes via a separate FD.
